Updated alignments to allow robust alignment of longer reads
Altered the primer stripping so that matches can occur anywhere within
the primer. This allows primers to be stripped even when there were
sequencing errors.

Altered bowtie alignments to be based on number of mismatches within
the first 28 nucleotides. Also, there is an option to set the maximum
combined Phred score for all mismatches.

Updated oLSC007.fna.

Changed some of the default parameters and edited the help statements
to describe what the default parameters area.